Combined CompositeCommands into 1 class
